Added publishing scripts
