+++ .nvmrc
@@ -1,0 +1,1 @@
+6.9.1
