+++ .travis.yml
@@ -1,0 +1,22 @@
+language: node_js
+
+addons:
+  firefox: "52.0.2"
+
+branches:
+  only:
+    - master
+    - /^[0-9]+\.[0-9]+\.[0-9]+.*/
+
+before_script:
+  - "export DISPLAY=:99.0"
+  - "sh -e /etc/init.d/xvfb start"
+  - sleep 3 # give xvfb some time to start
+
+script: npm run ci
+
+after_success:
+  - bash <(curl -s https://codecov.io/bash)
+  - node ./scripts/prepare-package
+  - cd dist
+  - bash <(curl -s https://blackbaud.github.io/skyux-travis/after-success.sh)
