+++ README.md
@@ -1,7 +1,13 @@
 # Help Widget Client
-## A SKY UX Builder client module to allow SPA's to consume the Help Widget
+
+
+[![npm](https://img.shields.io/npm/v/@blackbaud/help-client.svg)](https://www.npmjs.com/package/@blackbaud/help-client)
+[![status](https://travis-ci.org/blackbaud/help-client.svg?branch=master)](https://travis-ci.org/blackbaud/help-client)
+
+Provides a client-side library for interacting with the Help Widget. This module is a dependency of [SKY UX Builder](https://github.com/blackbaud/skyux-builder).
 
 ### skyux-config.json
+
 ```
 {
   "help": {
@@ -9,4 +15,5 @@
   }
 }
 ```
+
 Learn more about the Help Widget's [configuration options](https://github.com/blackbaud/bb-help).
