+++ package.json
@@ -1,7 +1,8 @@
 {
   "name": "@blackbaud/help-client",
   "version": "1.0.0",
-  "description": "",
+  "description": "Provides a client-side library for interacting with the Help Widget.",
+  "module": "index.ts",
   "engines": {
     "node": ">=6.3.1"
   },
@@ -10,11 +11,19 @@
     "url": "https://github.com/blackbaud/help-client.git"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "lint": "tslint 'src/**/*.ts'",
+    "test": "npm run lint"
   },
   "keywords": [],
-  "author": "Blackbaud",
-  "license": "ISC",
+  "author": "Blackbaud, Inc.",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/blackbaud/help-client/issues"
+  },
+  "homepage": "https://github.com/blackbaud/help-client#readme",
   "dependencies": {},
-  "devDependencies": {}
+  "devDependencies": {
+    "tslint": "5.1.0",
+    "typescript": "2.2.2"
+  }
 }
