+++ tslint.json
@@ -1,0 +1,28 @@
+{
+  "extends": [
+    "tslint:recommended"
+  ],
+  "rules": {
+    "member-ordering": [
+      true,
+      "public-before-private",
+      "static-before-instance",
+      "variables-before-functions"
+    ],
+    "no-console": [false],
+    "no-string-literal": false,
+    "quotemark": [
+      true,
+      "single",
+      "avoid-escape"
+    ],
+    "switch-default": false,
+    "trailing-comma": [
+      true,
+      {
+        "multiline": "never",
+        "singleline": "never"
+      }
+    ]
+  }
+}
